first instrumentation performance comparison test between new and old list impls.
